move spell cost computation to spellbook
