+++ game/magic/player/player_test.go
@@ -2,7 +2,6 @@ package player
 
 import (
     "testing"
-    "math"
 
     "github.com/kazzmir/master-of-magic/game/magic/hero"
     "github.com/kazzmir/master-of-magic/game/magic/setup"
@@ -79,72 +78,4 @@ func TestResearchPoints(test *testing.T) {
     if computeEffectiveResearchPerTurn(&wizard, points, natureSpell) != int(points * (1 + 0.1 + 0.25)) {
         test.Errorf("Research points computation doesn't work")
     }
-}
-
-func TestSpellCost(test *testing.T) {
-    natureSpell := spellbook.Spell{
-        CastCost: 50,
-        Magic: data.NatureMagic,
-        Section: spellbook.SectionSpecial,
-    }
-
-    wizard := setup.WizardCustom{
-        Books: []data.WizardBook{
-            data.WizardBook{
-                Magic: data.NatureMagic,
-                Count: 1,
-            },
-        },
-        Retorts: []data.Retort{
-        },
-    }
-
-    value := computeEffectiveSpellCost(&wizard, natureSpell, true, false)
-    if value != 50*5 {
-        test.Errorf("casting cost for nature spell with no modifiers was wrong. expected=%v actual=%v", 50*5, value)
-    }
-
-    // 30% reduction
-    wizard = setup.WizardCustom{
-        Books: []data.WizardBook{
-            data.WizardBook{
-                Magic: data.NatureMagic,
-                Count: 10,
-            },
-        },
-        Retorts: []data.Retort{
-        },
-    }
-
-    value = computeEffectiveSpellCost(&wizard, natureSpell, true, false)
-    if value != 50*5*7/10 {
-        test.Errorf("casting cost for nature spell was wrong. expected=%v actual=%v", 50*5*7/10, value)
-    }
-
-    // 30% reduction + nature mastery
-    wizard = setup.WizardCustom{
-        Books: []data.WizardBook{
-            data.WizardBook{
-                Magic: data.NatureMagic,
-                Count: 10,
-            },
-        },
-        Retorts: []data.Retort{
-            data.RetortNatureMastery,
-        },
-    }
-
-    value = computeEffectiveSpellCost(&wizard, natureSpell, true, false)
-    expected := 50*5*(100 - (30 + 15))/100
-    if value != expected {
-        test.Errorf("casting cost for nature spell was wrong. expected=%v actual=%v", expected, value)
-    }
-
-    // evil omens
-    value = computeEffectiveSpellCost(&wizard, natureSpell, true, true)
-    expected = int(math.Floor(50.0*5*(100 - (30 + 15))/100 * 3/2))
-    if value != expected {
-        test.Errorf("casting cost for nature spell with evil omens. expected=%v actual=%v", expected, value)
-    }
-
 }
